+++ src/main/java/com/telepathicgrunt/blame/Blame.java
@@ -9,7 +9,7 @@ public class Blame implements ModInitializer
 {
     public static final String MODID = "blame";
     public static final Logger LOGGER = LogManager.getLogger();
-    public static String VERSION = "2.1.2";
+    public static String VERSION = "2.2.0";
 
     @Override
     public void onInitialize() {